Abstract Split over Collection instead of Array
