+++ Package.swift
@@ -18,7 +18,7 @@ let package = Package(
         // Sources
         .target(name: "Destructure"),
         .target(name: "Algebra", dependencies: ["Destructure"]),
-        .target(name: "DataStructures", dependencies: ["Algebra", "Destructure"]),
+        .target(name: "DataStructures", dependencies: ["Destructure", "Algebra", "Algorithms"]),
         .target(name: "Algorithms", dependencies: ["Destructure"]),
 
         // Tests
