+++ Sources/Algorithms/Split.swift
@@ -1,0 +1,38 @@
+//
+//  Split.swift
+//  Algorithms
+//
+//  Created by James Bean on 7/11/18.
+//
+
+#warning("Get rid of constraint of Index == Int")
+extension Collection {
+
+    /// - Returns: A two-tuple containing two arrays of `Elements` split at the given `index`,
+    /// if the given `index` is in the bounds of `self.` Otherwise, `nil`.
+    //
+    // FIXME: Generalize to `Collection`
+    // FIXME: Return (SubSequence,SubSequence)?
+    public func split(at index: Index) -> ([Element], [Element])? {
+        guard index >= startIndex && index <= endIndex else { return nil }
+        let left = Array(self[startIndex ..< index])
+        let right = index == endIndex ? [] : Array(self[index ..< endIndex])
+        return (left, right)
+    }
+
+    /// - returns: A three-tuple containing:
+    ///
+    /// - The elements to the left of the element at the given `index`
+    /// - The element at the given `index`
+    /// - The elements to the right of the element at the given `index`
+    ///
+    /// if the given `index` is in the bounds of `self.` Otherwise, `nil`.
+    public func splitAndExtractElement(at index: Index) -> ([Element], Element, [Element])? {
+        guard count > 0 else { return nil }
+        guard index >= startIndex && index < endIndex else { return nil }
+        let element = self[index]
+        let left = Array(self[startIndex ..< index])
+        let right = index == endIndex ? [] : Array(self[self.index(after: index) ..< endIndex])
+        return (left, element, right)
+    }
+}
