+++ Sources/DataStructures/Tree.swift
@@ -7,6 +7,7 @@
 //
 
 import Algebra
+import Algorithms
 
 /// Value-semantic, immutable Tree structure.
 public enum Tree <Branch,Leaf> {
